+++ opening-hours-syntax/src/lib.rs
@@ -1,4 +1,8 @@
 #![doc = include_str!("../README.md")]
+#![cfg_attr(not(feature = "std"), no_std)]
+
+#[macro_use]
+extern crate alloc;
 
 #[macro_use]
 extern crate pest_derive;