+++ scripts/check-version.py
@@ -48,7 +48,7 @@ async def main():
     sy_version = toml.load(rt / "opening-hours-syntax/Cargo.toml")["package"]["version"]
     cc_version = toml.load(rt / "compact-calendar/Cargo.toml")["package"]["version"]
     py_version = toml.load(rt / "opening-hours-py/Cargo.toml")["package"]["version"]
-    pt_version = toml.load(rt / "pyproject.toml")["tool"]["poetry"]["version"]
+    pt_version = toml.load(rt / "pyproject.toml")["project"]["version"]
 
     print("Checking local packages:")
     print(" - Rust crate:", rs_version)